Modified the "nearby" computation
Now only replies if the requesting stop is in our immediate future, not our arbitrary future. This is because the RouteInfoRequest messages are expected to be propagated one "hop" at a time from stop to stop, so we don't need to respond when we aren't nearby; if we're the closest bus, one of our immediate neighbor stops will eventually find us.